scripts/eval_scripts_tifa&genai/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.37.0,<5
accelerate>=0.25.0
diffusers>=0.25.0
peft>=0.7.0
bitsandbytes>=0.41.0
janus @ git+https://github.com/deepseek-ai/Janus.git
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.66.0
openai>=1.6.0
pandas>=2.0.0
pytest>=7.4.0

